requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0